Move Header, Card to their own directory
