+++ src/components/Post/PostList.js
@@ -2,6 +2,7 @@ import React from 'react'
 import Post from './Post'
 import { useQuery } from '@apollo/client'
 import { POSTS_QUERY } from '../../graphql/queries/POSTS_QUERY';
+import Loader from '../UI/Loader/Loader';
 
 const PostList = () => {
   const { data, loading } = useQuery(POSTS_QUERY)
@@ -15,6 +16,9 @@ const PostList = () => {
           ))}
         </section>
       ) }
+
+
+      <Loader loading={loading} data={data} />
     </>
   )
 }
